+++ tests/test_update_attributes.py
@@ -4,7 +4,7 @@ import unittest
 
 import wrapt
 
-from wrapt import six
+from compat import PY2, PY3, exec_
 
 @wrapt.decorator
 def passthru_decorator(wrapped, instance, args, kwargs):
@@ -45,7 +45,7 @@ class TestUpdateAttributes(unittest.TestCase):
         def function():
             pass
 
-        if six.PY3:
+        if PY3:
             method = self.test_update_qualname
             self.assertEqual(function.__qualname__,
                     (method.__qualname__ + '.<locals>.function'))
@@ -63,7 +63,7 @@ class TestUpdateAttributes(unittest.TestCase):
 
         instance = wrapt.FunctionWrapper(function, wrapper)
 
-        if six.PY3:
+        if PY3:
             method = self.test_update_qualname_modified_on_original
             self.assertEqual(instance.__qualname__,
                     (method.__qualname__ + '.<locals>.function'))
@@ -134,7 +134,7 @@ class TestUpdateAttributes(unittest.TestCase):
         def function():
             pass
 
-        if six.PY3:
+        if PY3:
             self.assertEqual(function.__annotations__, {})
 
         else:
@@ -157,7 +157,7 @@ class TestUpdateAttributes(unittest.TestCase):
 
         instance = wrapt.FunctionWrapper(function, wrapper)
 
-        if six.PY3:
+        if PY3:
             self.assertEqual(instance.__annotations__, {})
 
         else: