+++ docs/changes.rst
@@ -1,6 +1,40 @@
 Release Notes
 =============
 
+Version 1.9.0
+-------------
+
+**Features Changed**
+
+* When using ``wrapt.wrap_object()``, it is now possible to pass an
+  arbitrary object in addition to a module object, or a string name
+  identifying a module. Similar for underlying ``wrapt.resolve_path()``
+  function.
+
+**Bugs Fixed**
+
+* It is necessary to proxy the special ``__weakref__`` attribute in the
+  pure Python object proxy else using ``inspect.getmembers()`` on a
+  decorator class will fail.
+
+* The ``FunctionWrapper`` class was not passing through the instance
+  correctly to the wrapper function when it was applied to a method of an
+  existing instance of a class.
+
+* The ``FunctionWrapper`` was not always working when applied around a
+  method of a class type by accessing the method to be wrapped using
+  ``getattr()``. Instead it is necessary to access the original unbound
+  method from the class ``__dict__``. Updated the ``FunctionWrapper`` to
+  work better in such situations, but also modify ``resolve_path()`` to
+  always grab the class method from the class ``__dict__`` when wrapping
+  methods using ``wrapt.wrap_object()`` so wrapping is more predictable.
+  When doing monkey patching ``wrapt.wrap_object()`` should always be
+  used to ensure correct operation.
+
+* The ``AttributeWrapper`` class used internally to the function
+  ``wrap_object_attribute()`` had wrongly named the ``__delete__`` method
+  for the descriptor as ``__del__``.
+
 Version 1.8.0
 -------------